requirements.txt
fastnumbers>=3.1.0
Flask>=2.0.1
flask-mongoengine>=1.0.0
mimerender-pr36>=0.0.2
mongoengine>=0.23.1
pymongo>=3.12.0
python-dateutil>=2.8.2
unflatten>=0.1
